Time: 99 ms (69.09%), Space: 130.6 MB (48.52%) - LeetHub

+++ 0134-gas-station/0134-gas-station.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int canCompleteCircuit(vector<int>& gas, vector<int>& cost) {
-        long tg = 0,tc = 0;
+        int tg = 0,tc = 0;
         for(int i=0;i<gas.size();i++){
             tg+=gas[i];
             tc+=cost[i];
